pyproject.toml
[project]
name = "finance-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "chekh",email = "chechirkinbogd@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
